Strava Heatmap
This feature is in beta! WIP! It uses my heatmap cloudfront policies for high-res, do know that they expire every week, and I use them for testing, you will need to use your own credentials for high-res tiles! In case you want to use low-res, it is provided in the comment